Move odometry display and covariance classes
