+++ rviz_common/src/rviz_common/covariance_visual.cpp
