+++ rviz_default_plugins/src/rviz_default_plugins/displays/odometry/odometry_display.cpp
@@ -38,7 +38,7 @@
 #include <OgreSceneManager.h>
 #include <OgreSceneNode.h>
 
-#include "odometry_display.h"
+#include "odometry_display.hpp"
 #include "covariance_property.h"
 #include "covariance_visual.h"
